Use EventTarget shim instead of the one in Window.
`window` in the Firefox privileged code context refers to the `hiddenDomWindow`.
On Windows and Linux, this window is loaded with a unpriviledged principal
instead of the all-powerfull system one. On these platforms and because of X-Ray
shenanigans we get errors when subclassing EventTarget.
Therefore we simply ship our own version of EventTarget.